src/instagram-crawler/python/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
instaloader==4.9.6
python-dotenv==0.21.1
requests==2.28.2
urllib3==1.26.14